% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapstyles-package.R
\docType{package}
\name{mapstyles-package}
\alias{mapstyles}
\alias{mapstyles-package}
\title{mapstyles: Modern Map Styles Toolkit}
\description{
Provides an experiemental S7 object model for constructing, validating, converting, and managing Maplibre Style Specification documents.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jimbrig/mapstyles}
  \item Report bugs at \url{https://github.com/jimbrig/mapstyles/issues}
}

}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@jimbrig.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})

Authors:
\itemize{
  \item Jimmy Briggs \email{jimmy.briggs@jimbrig.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})
}

}
\keyword{internal}
